Fix edgedone when a restat rule updates a file
In development, the code was structured to have two loops, one to
prune dependent actions if output didn't change, and one to unblock
dependent edges if it did. However, with a single loop, we still
need to call nodedone if we determine that we can't prune an action
so that it is also unblocked.

This fixes a bug where actions dependent on a restat rule wouldn't
run if the rule actually updated the file.