Fix order-only input handling in some cases
Previously, an action with an order-only input would only block actions
that depended on it directly. However, if the dependency is indirect, it
still needs to block the action. To do this, increase nblock for any
input that is blocked (or dirty).

Now that a clean action can block a dependent, when a target is rebuilt
we also need to consider dependent actions that are *not* dirty (since
they may be along the path to the dirty action). We can treat these just
like actions skipped due to restat.