Allocate n->use as needed
We only need to populate it for dependent edges that are blocked by that
node. If everything is clean, we don't need it at all.

Also, just allocate with increasing powers of two. Although we know the
upper-bound from reading the build manifest, the $depfile may add
additional dependent edges that we can't anticipate.