parse: Make multiple generating actions a warning instead of an error
Although this almost always indicates a bug in the build system, it
appears to be widespread enough[0] to warrant matching ninja's default
behavior here.

Ninja handles this case by effectively pretending that the output was not
listed in any subsequent actions. This may cause issues if the subsequent
action gets invoked to build some other output and updates the duplicate
output, or if the command being run depends on some expected number of
file arguments from the expansion of $out. However, it does seems like
a sane way to handle a non-sensical situation, and does not introduce
any complexity in the main build logic.

So for now, just emit a warning and skip over the output like ninja does.

[0] https://marc.info/?l=openbsd-ports&m=153737668917497&w=2