build: Fix isdirty for restat=1, generator=1, and no build log entry
This was accidentally broken in d6c8545ca8, which fixed isdirty for
generator=1 and no build log, but did not consider restat=1. The
"older than input" check was skipped for restat=1 because a newer
log entry can override an old output (due to restat detecting no
change). Instead, it relied on the "missing log entry" check.
However, when d6c8545ca8 fixed that check to ignore generator
actions, it had the side effect of failing to mark actions with
restat and generator as dirty even if the output was older than the
newest input.