env: Fix variable lookup order between rule and parent environment
The documented behavior at https://ninja-build.org/manual.html is:

1. Special built-in variables ($in, $out).
2. Build-level variables from the build block.
3. Rule-level variables from the rule block (i.e. $command). (Note
   from the above discussion on expansion that these are expanded
   "late", and may make use of in-scope bindings like $in.)
4. File-level variables from the file that the build line was in.
5. Variables from the file that included that file using the subninja
   keyword.

Due to https://github.com/ninja-build/ninja/issues/1516, ninja's
actual behavior depends on whether or not the edge has any bindings.
If it does, it behaves as 1, 2, 3, 4, 5 as documented. If it doesn't,
it behaves as 1, (2 empty), 4, 3, 5.

Previously, we looked up a variable first in the edge's bindings,
then the file's bindings, and then the rule's bindings. This is 1,
2, 4, 5, 3, which roughly corresponded to ninja's behavior for edges
with no bindings.

Rather than try to match ninja's quirk, just implement the documented
order.