deps: Fix depfile parsing with multiple `output: input` lines
The `if (sawcolon)` branch ends with the next character already in `c`,
but the `else` branch still has the last character it consumed (`:`).

The following loop reads a new character at the beginning of each
iteration, so if the next character was already in `c`, it would get
discarded.

To resolve this, make the `else` branch match the `if (sawcolon)` branch
by reading the next character, then adjust the following loop to read
the next character at the end of each iteration.

Fixes #24.