Fix manifest rebuild when it was pruned due to restat
If build.ninja depends on an output with `restat = 1`, it may get
pruned if the output did not change. In this case, we can continue
with the build, since the previously loaded manifest is still
correct.

To make sure that we can add new build targets and start a new build
after executing the build for another target, add a function to
reset the build state.

Fixes #31.