add jobserver protocol
GNU Make has a neat feature called the jobserver protocol, where
the top-level Make can allocate a specific number of job slots, and
child makes can take slots to do work in. This was designed to stop the
parallelisation problem where a top-level make -j10 may potentially spawn
10 separate sub-makes all with -j10 so there's now 100 parallel jobs.

However, it's also useful for resource control in systems which build
multiple pieces of software at once. For example, Bitbake can build
N different pieces of software at once, and each of those is passed a
-jM flag. If each of these N tasks is compiling then thats's N*M jobs
so you don't want N or M to be too high, but if only 1 of N is building
then you want M to be high.

With the job server protocol there are N slots in total for all sub
makes, so you can control the resource utilisation more accurately. By
supporting the jobserver protocol instead of just -j, Samurai can join
in the resource pooling and builds can be more efficient.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>